Simplified package structure.
